The first working version!
This is the first working version of the fireworks simulator! It makes a single red firework in the center of the screen. Noteable bugs: the particles are supposed to be brightest in the center and fade darker, but they are instead perfect circles. Also, gravity was coded, but does not seem to be working at all. The next thing to do would be to fix those bugs.