I fixed angle distributions and center mass.
First, angle distributions were not distributed evenly as they should have been. I fixed that by realizing that sin and cos take radians, not degrees, and using PI instead of 90s. This is why programming languages need units!!
It was also ugly to see the mass of particles that just sit at the center. I realized that this is because the radius of the first circle on which particles lie was 0, so I made the radii start at 1 and go up to the max radius instead of starting at 0 and going to the max radius - 1.
Next up: gravity!