Gravity!
I increased the gravity factor, refersed the sign (because I forgot that negative is down in GLSL), and made the initial velocity calculation depend on the circle the particle lay on (as it should). Now gravity works! If I made it 3-D, it could make the nice curved paths seen in real fireworks; this is what a firework would look like if it exploded on a perfectly vertical plane.